service_organization/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.24.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.0
Deprecated==1.2.13
ecdsa==0.17.0
exceptiongroup==1.0.4
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
humanize==4.3.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
path==16.4.0
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pyzbar==0.1.9
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a23
sqlmodel==0.0.6
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.17.6
fastapi-cache2[redis]==0.2.1
xlrd==2.0.1
openpyxl==3.1.2
openpyexcel==2.5.14
